requirements-mac.txt
Flask
numpy
Pillow
requests
tensorflow-macos==2.9.0
